Support for typechecking hooks and filters
As well as supporting stubs for all WordPress functions / classes, we also want an index of all filters / actions. From the type systems point of view, `add_filter` / `add_action` are quite simple, we can improve that a lot to make sure all WordPress hooks are type sound. This can catch a lot of common mistakes with hooks (like wrong types expected in the function, forgetting to return the correct type, or forgetting to the `passed params` number to `add_filter` etc.

This also includes an analyzer for customer filters in the current codebase, so those can be type checked too.